+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [0.5.2] (2026-04-16)
+
+### Improvements
+
+- Add `data-` attribute to strengthen month index traction in `Calendar`.
+
+
+### Tests
+
+- Set up test `vitest` testing environment.
+- Add test suites for the following components:
+  - `Button`.
+  - `Container`.
+  - `ArticlesNav`.
+  - `BottomNavbar`.
+  - `CheckboxFormGroup`.
+  - `DefaultFormGroup`.
+  - `FilterTogglerCheckboxBtn`.
+  - `Footer`.
+  - `Head`.
+  - `Calendar`.
+
+
+### Bug Fixes
+
+- Remove accidental `generateHexCode()` call in GET `auth/session`.
+- Added missing day header in `Calendar`.
+
+
 ## [0.5.1] (2026-04-11)
 
 ### Changes